Refacting the Lua-C code to remove common structs and functions to a new private header and file. This reduces duplication between the various App Lua-C wrappers.
