+++ apps/lw_priv.c
@@ -1,0 +1,33 @@
+#ifdef GKYL_HAVE_LUA
+
+#include <gkyl_lw_priv.h>
+
+void
+gkyl_lw_eval_cb(double t, const double* GKYL_RESTRICT xn, double* GKYL_RESTRICT fout, void *ctx)
+{
+  struct lua_func_ctx *fr = ctx;
+  lua_State *L = fr->L;
+
+  int ndim = fr->ndim;
+  int nret = fr->nret;
+  lua_rawgeti(L, LUA_REGISTRYINDEX, fr->func_ref);
+  lua_pushnumber(L, t);
+  lua_createtable(L, GKYL_MAX_DIM, 0);
+
+  for (int i=0; i<ndim; ++i) {
+    lua_pushnumber(L, xn[i]);
+    lua_rawseti(L, -2, i+1); 
+  }
+
+  if (lua_pcall(L, 2, nret, 0)) {
+    const char* ret = lua_tostring(L, -1);
+    luaL_error(L, "*** eval_ic ERROR: %s\n", ret);
+  }
+
+  for (int i=nret-1; i>=0; --i) { // need to fetch in reverse order
+    fout[i] = lua_tonumber(L, -1);
+    lua_pop(L, 1);
+  }  
+}
+
+#endif
